+++ optimizer.h
@@ -23,7 +23,7 @@
 struct optimizer_s;
 struct cpu_s;
 
-extern void cpu_opt(uint8_t, uint32_t, int8_t, linepos_t);
+extern void cpu_opt(uint8_t, uint32_t, int, linepos_t);
 extern void cpu_opt_invalidate(void);
 extern void cpu_opt_destroy(struct optimizer_s *);
 extern void cpu_opt_long_branch(uint16_t);
